+++ src/main/java/nextstep/ladder/domain/alternative/Direction.java
