+++ src/main/java/nextstep/ladder/domain/alternative/Point.java
