+++ src/main/java/nextstep/ladder/utils/Lists.java
@@ -1,0 +1,9 @@
+package nextstep.ladder.utils;
+
+import java.util.List;
+
+public class Lists {
+    public static <T> T last(List<T> list) {
+        return list.get(list.size() - 1);
+    }
+}
